requirements.txt
# conda create --name nlg python=3.6.1
tqdm==4.19.1
numpy==1.16.0
pandas==0.22.0
sklearn
gensim==3.7.0
nltk==3.2.3
Keras==2.0.8
tensorflow-gpu==1.3.0
matplotlib
jupyter